mostly refactored SketchException usage to a relocated processing.utils.SketchException, except I'm missing something because on compile it's not finding it. (module / package visibility problem or somethin')
