+++ java/src/processing/mode/java/preproc/PdePreprocessor.java
@@ -32,6 +32,7 @@ import org.antlr.v4.runtime.*;
 import org.antlr.v4.runtime.tree.ParseTree;
 import org.antlr.v4.runtime.tree.ParseTreeWalker;
 import processing.app.Preferences;
+import processing.utils.SketchException;
 
 
 /**
