+++ java/test/processing/mode/java/ParserTests.java
@@ -17,7 +17,7 @@ import org.junit.Test;
 import processing.app.Preferences;
 import processing.mode.java.preproc.PreprocessorResult;
 import processing.mode.java.preproc.PdePreprocessIssueException;
-import processing.mode.java.preproc.SketchException;
+import processing.utils.SketchException;
 
 
 public class ParserTests {
